Add and update checkout integration tests
These tests make actual SF API calls and verify that the response
returned from the SDK is what we expect. Our goal here is to ensure
that when a given checkout method is called with SDK v3, the response
is the same as what previous versions of the SDK returned (aside from
known/documented discrepancies between the cart/checkout APIs).

In the tests, the "expected" values are all values that we got when
we ran these operations with the previous SDK version. The "actual"
values are what v3 of the SDK returns for the same operation.

Co-authored-by: Juan P. Prieto <jp@calltheguys.co>